IDEA-325438 Fix position-above-caret logic in LookupUi
It's a very elusive and old bug, probably originating from old times
when most people used only one monitor and its Y coordinate of
the top edge was always zero.

Prior to this fix, the code that decided whether to position the popup
above the line was basically this:

shiftLow < 0 && shiftLow < location.y - dim.height && location.y >= dim.height

Here, shiftLow is (screen bottom Y minus popup bottom Y), so the first part
can be rewritten as

popup bottom Y > screen bottom Y

in other words, the popup doesn't fit below the caret. This is correct.

The second part is really hard to understand, but it was made redundant
after the third part was added. Indeed, if the first and the third parts are true,
then the second will always be true because it can be rewritten as

location.y > dim.height + shiftLow

The third part tells us that location.y >= dim.height and if we add a negative
number (shiftLow, according to the first part) to the right side, then we
certainly turn >= into >. So the second part is completely redundant.

And as for the third part, that's where the bug was. Even two bugs.

First, it should be checking that there's enough space above the caret for
the whole popup, where location.y is the amount of space and dim.height
is the hight of the popup. Well, the amount of space is really
location.y minus screenRectangle.y, so the code only worked
if the bottom of the screen is at y=0, which isn't always true
for multi-screen configurations.

Second, it didn't take the line height into account at all. It assumed
that if the popup would be placed above, its coordinate would be
location.y - dim.height, when actually it should be minus lineHeight as well.

The bug was elusive and remained unfixed for years, because an incorrect
calculation didn't always lead to a wrong placement. It would just
place the popup outside the screen, but then a lot of complicated logic
would kick in afterward, and in some cases that logic would place
the popup in really weird ways, like on another monitor (because it
happened to be closer to the wronly calculated location). That logic
is really intended as the last chance fallback, so at least the popup
doesn't show up outside of the visible area or partially offscreen,
so it shouldn't be relied upon for correct placement.

Fix by rewriting this logic using more self-documenting variable names.
Also, calculate the popup location in case of above the caret placement
in advance, and reuse it for both this check (will it fit?)
and the actual placement, so if that calculation ever changes, there
won't be any discrepancy between the check and the placement, like
it was here with not taking the line height into account.

GitOrigin-RevId: 3f884c18c7f6d7304abc30844cd7c8db82fd3548