[vcs-log] refactor isIndexingPaused function
The original function returned true when indexing was not scheduled, which didn't necessarily mean that indexing was paused. The new implementation accurately determines if indexing is paused by checking if project repositories are present in the VcsLogBigRepositoriesList. The previous version of the function was inlined and the behavior of the using code was not altered.

GitOrigin-RevId: ba2762a8749ef678815b1cad47215da2bb0ebdfd