EA-836208 Fix already disposed error in FileStructurePopup.showWithResult
Using onProcessed() here caused it to be invoked if the promise is cancelled
because, for example, Esc is pressed before the tree was fully built.
In that case an attempt to complete the initialization would sometimes
cause an exception. Since it doesn't make any sense to initialize
anything if it was cancelled anyway, simply use onSuccess()
instead of onProcessed().

GitOrigin-RevId: 1ed166745215e75d61b97f7e63f61821ea5c3a17