Remove deprecated method getAssociatedExtensions from FileTypeManager
The method getAssociatedExtensions was removed from the FileTypeManager, FileTypeManagerImpl and MockFileTypeManager. These changes were made as a result of the function becoming obsolete, as file type associations are not only limited to the extension list - they are now obtained from the getAssociations method. This helps to consolidate function calls and reduce redundancy in the codebase.

GitOrigin-RevId: c59b8a21f7347723e4c9ca0ef0dbadcb27aac2ef