IDEA-309322, EA-772915 Move image loading out of EDT for ImageEditorImpl
Create a service and use coroutines to freely jump between IO threads and EDT.
To avoid converting too much stuff to Kotlin at once, use dependency
inversion and provide an object for image file loading to ImageEditorImpl
through the new service.

Cancelling the image loader's job through dispose() looks a bit ugly,
but until we move the entire thing to coroutines, it seems to be the only reasonable
way.

GitOrigin-RevId: ce42805acfcb67bbdbb3de869fb7a3606adcf383