[Kotlin] Update testdata to make tests pass
Because of the previous commits (KT-59304) the HMPP is enabled by default if the corresponding property
is not specified. And for very old projects it is not specified (here in test 1.3.61 is used).

That slightly changed the test behaviour. But the fact that we are still testing something against 1.3.61
is something that we should solve.

#KTIJ-26437 Open

GitOrigin-RevId: f3c4fb290d728308b34aada7d95d1335e7d8c7c4