[Workspace Model] [IDEA-326804] Throw `rootsChanged` not on any change of `ModuleEntity` but on change of dependencies
The current implementation of `rootsChanged` throws an event on any modification of tracked entity and not on change of used field.
Together with IDEA-313747 this causes a problem that some approach for creation of a `FacetEntity` may throw a `rootsChanged` what breaks `FacetManagerTest#testAddDeleteFacet` contract.
With this change, we throw an event only when the "dependencies" field of `ModuleEntity` was modified. This solution can be improved by not storing such logic in a separate function. Alternatively, the new API IDEA-318790 should solve the issue when there is a split of what fields are read and when we should react on change.

GitOrigin-RevId: 4ab47f094dd54eb95a6b4e7a85f935ebd4d829a7