Fix import of Gradle projects after breaking change in compiler arguments parsing logic
As a side effect of fixing KT-58183 (ParseCommandLineArgumentsKt.parseCommandLineArguments takes ~500ms) the import functionality was broken due to incompatibility of annotation filtering process. Old approach relied on annotation of Kotlin-properties, but the new one is based on java reflection with annotations on java-fields, that gives a gain in time. This commit supports new approach based on java reflection of annotation on fields.

(#KT-60308, #KT-60195) Fixed

GitOrigin-RevId: 3e2bda965cc48589346205fcec94cd44009e2990