IDEA-309322, EA-830243 Create ImageOrColorPreviewService to get rid of slow ops
Create a service and use coroutines to freely jump between IO threads and EDT.

Implementation note: to make sure the service always detaches from the editor
and not leaks anything, the whole attach/detach stuff is placed into a single
coroutine with a try-finally block with the finally block wrapped into
a non-cancellable context to make sure it always completes.
This way, whether the coroutine scope is cancelled or the editor is released
(the job is cancelled on such a release), it'll always detach.

GitOrigin-RevId: 91501506eb3aafd9d4a267602dab53116bc1e638