IDEA-309322, EA-830243 Get rid of remaining slow ops in ImageOrColorPreviewService
Now that we have a Kotlin service with a coroutine scope,
just split the remaining code into read actions and EDT code
to get rid of existing slow ops marked with allowSlowOperations.

The code checking for "displayed elements == elements under cursor"
had to be moved from the mouse handler (non-suspending context)
to HidePreviewRequest.

GitOrigin-RevId: d79d3d836fa532427c626a7e390b279201e3e375