IDEA-314590 Implement factory default tool window layout
Use the empty string as its name because it's an invalid name
for a custom layout.

The Store Layout action is disabled when the default layout is active
because it's read-only. However, if someone tries to do the same
thing through the API (ToolWindowDefaultLayoutManager.setLayout),
as Rider does, for example, then we instead overwrite the "Custom"
layout and make it the current one. This isn't the best thing to do,
but it'll have to work until Rider implements multiple tool window layouts
on its side.

The Restore Current Layout action works the same way as usual
even if the currently active layout is the factory default one,
we only substitute the description because there's no name to use.

GitOrigin-RevId: 36f6f98bb4d34e4bc613088c7a1c0487be62c98f