IDEA-327037 don't treat unstarted progress as stopping
This change reverts 30f4092027995771d7a0a4b97552d29f1b5808bd, and adds another fix for coroutine-based progresses.

Calling `indicator.start` before adding in the UI resulted in `InlineProgressIndicator.wasStarted` remaining `false`,
which is turn didn't display "Stopping - " prefix for cancelled coroutine progresses. `indicator.start` should be invoked _after_ the progress is shown in the UI, because when it's added to the UI
`InfoAndProgressPanel` registers a bunch of state delegates, one of which tracks `indicator.start`.

With this fix in place, 30f4092027995771d7a0a4b97552d29f1b5808bd is not needed anymore as it caused IDEA-327037,
and probably causes a bunch of other problems we are unaware of yet, so it's safer to keep the old behaviour.

GitOrigin-RevId: e37c0fff9258b89f7c293424daaa282c88a51b30