IDEA-326304 Use CoroutineStart.UNDISPATCHED in SelectInProjectViewImpl
The regression was caused by the fact that when the service is first used,
then it's initialized and tryEmit is called immediately, which happens
before the collectLatest is even dispatched. This is exactly why it almost
never worked on the first call.

By using UNDISPATCHED, we guarantee that collectLatest is executed
until the first suspension point before init completes, and therefore
before any event is emitted.

GitOrigin-RevId: 41cbcae73c8c23f4b6b46e264c7c90fd12fd3d44