[rdct] update UI on-the-fly when UI settings are changed in another process (RDCT-366)
When multiple JetBrains Client processes (RDCT-6) are running in parallel, changes in one of them will cause changes in configuration files which will be applied to all the other processes. So even changes in NotRoamableUiOptions must cause 'uiSettingsChanged' event to be fired.

Code which calls LafManager.updateUI was moved from AppearanceConfigurable to LafManagerImpl to ensure LaF is updated after such changes.

GitOrigin-RevId: b8a46d250d343590a250b83e5f73b389bebcb91d