[kotlin-dfa] Remove flaky test with two boolean contracts
It looks like some rare race happens when two contracts are present: sometimes callDescriptor.getUserData(ContractProviderKey)?.getContractDescription() returns single effect, even though two are declared. Was not able to reproduce it in isolated test. Given that the contracts are experimental, and no stdlib method has two contracts, I think we can postpone the investigation. In any case, it's not the inspection defect. Probably it won't happen in K2 anyway

GitOrigin-RevId: fbab15a68928c06192a11c97770f70557feac294