IDEA-327188 Properly set target graphics device for Ctrl+hover editor hints
The root cause of the bug is that for some reason this component
(created by com.intellij.ide.IdeTooltipManager.initPane and wrapped
into a HintUtil.HintLabel) doesn't update its preferred size properly
once it's attached to a real screen. And since its preferred size is
calculated before that
(in com.intellij.codeInsight.navigation.CtrlMouseHandlerKt.wrapInScrollPaneIfNeeded
to be precise), it defaults to the graphics configuration of the main
monitor (as per the OS settings) and then just sticks to that incorrect size.

The fix is more like an ugly hack, and it's local to CtrlMouseHandler,
which means other HintUtil.createInformationLabel usages may
still be affected. We'll need to think of a better way to do it
or, failing that, check other usages and apply the fix there.

Unfortunately, it doesn't seem possible to fix it deeper in the API
because there are no components passed in there, so we can't
get a graphics device.

GitOrigin-RevId: 2276724c0f16734fedd921166837187bbb1a2818