[project structure] don't create editor for all facets when a facet type node is selected in Project Structure dialog (IDEA-327311)
Before when a facet type node is selected under 'Facets' item in 'Project Structure' dialog, IDEA tried to create an instance of MultipleFacetSettingsEditor for all facets. This triggered synchronous initialization of editors for all facets of the selected type even if 'createMultipleConfigurationsEditor' isn't implemented for this FacetType, causing a UI freeze if there are many facets (IDEA-208594). Removal of AllFacetsConfigurable fixes the problem.

GitOrigin-RevId: 9f810f119abd45b35b46d6ccf8306066720b36e6