[android] IDEA-307652 - Enable Device Mirroring feature and package ffmpeg in IDEA distributive.
To enable the Device Mirroring feature in the production distribution,
we have bundled the ffmpeg and javacpp native binaries with the IDEA distribution.
These binaries are specific to each operating system and architecture, and they are packaged within jar files.

To prevent unnecessary inflation of the IDEA distribution size,
we only include ffmpeg and javacpp libraries that are compatible with the current operating system and architecture.

To accomplish this, we define separate ffmpeg and javacpp module libraries per operating system and architecture
in the intellij.android.streaming.iml module. Additionally, in the CommunityRepositoryModules,
we create an Android plugin layout for each operating system and architecture,
and we only package the corresponding ffmpeg and javacpp libraries within them.

GitOrigin-RevId: 619654511fc6d937c6d544d4409b24970afd1ff5