[vcs-log] run index diagnostic for roots with paused indexing
When indexing is paused for the root, information for the already processed commits is not removed. This data is used to display commit details in the table, so it is necessary to check it for correctness.

IDEA-327208

(cherry picked from commit 9c7c1c65cbed96721b7485b3e980c9e834d38068)

GitOrigin-RevId: 5eb3d4618b7f58653eefbc67857757ed0aa54a47