Convert ImageOrColorPreviewService to project-level to avoid leaks
Apparently, relying on detach() is not very good idea, as there
are leaks reported by tests. Let's see if using a project-level service
fixes it. But it's a good idea anyway, as previews are created
for a specific project anyway (attach just returns in project is null).

The only side effect is that now showing a preview for one project
won't cancel previews for other projects, but that's likely not a bad thing.

Now that it's guaranteed that the coroutine scope is cancelled
when the project is disposed, there's no need for that ugly
withContext(NonCancellable) in the finally block. Now the editors
are attached/detached as they are created and released, and on the
project dispose everything is cancelled and cleaned up anyway.

getService() is replaced with getServiceIfCreated() in editorReleased
because an editor can be released after the service is disposed.

GitOrigin-RevId: 0f9cf9c150cffdd05937a34e501ec0dbe2cc16e8