[kotlin] KTIJ-26376 Provide collective dependents for `KtLibraryModule`s based on `LibraryInfo`
Summary:

- If a `LibraryInfo` comprises multiple, deduplicated `LibraryEx`s, `IdeKotlinModuleDependentsProvider` only provided the dependents of the anchor library.
- The solution is to consider `LibraryInfo`s as a collection of `LibraryEx`s and to provide a union of *all* dependents. This solution leads to improved session invalidation consistency as well.

#### Problem

Consider the following `KtModule` structure with `M1` being a `KtLibraryModule` based on a `LibraryInfo` `I1` with anchor library `L2`:

```
       M1
     /    \
 L2 /      \ L3
   /        \
  M2        M3
```

Let's say `L3`'s roots are replaced and `KotlinModuleStateModificationService`'s workspace listener notices it. Then `invalidateLibraryModule` fetches `I1` from the library info cache and publishes a modification event for `M1`. Currently, the invalidation service will find only `M2` as a dependent, because `IdeKotlinModuleDependentsProvider` returns dependents only for the library info's anchor library. `M3` is not recognized as a dependent, even though the original dependency link via `L3` is from `M1` to `M3`.

An idea to fix this was to have multiple `LibraryInfo`s which are equal to each other, but contain the original `LibraryEx` instances with which they were created (so one library info based on `L2`, and another one based on `L3`). This would allow `KotlinModuleStateModificationService` to get only `M3` as a dependent of `L3`, because `LibraryInfoCache` would return a `LibraryInfo` for `L3` which contains `L3` as its anchor library instead of `L2`, while still being equal to the library info of `L2` (as we would want to keep the property that libraries with the same roots share the same session, for example).

However, it would not be correct to only invalidate `M3`'s session either. Because a modification event is published for `M1`, its session is cleared. Symbols from that session cannot only be referenced from `M3`'s session, but also from `M2`'s session. So effectively *both* sessions need to be invalidated.

This example illustrates the nature of `LibraryInfo`s well. As long as we want different `LibraryEx` instances with the same roots to share the same library sessions and have the same entries in caches (such as `LibraryDependenciesCache`), we have to treat such libraries as a *collection* and `LibraryInfo` is such a collection. A `KtLibraryModule` refers to a specific `LibraryInfo`, not a specific `LibraryEx`. The dependents computed by `IdeKotlinModuleDependentsProvider` must be the dependents of the `LibraryInfo`, not a single `LibraryEx`, and these are the dependents of *all* associated `LibraryEx` instances.

This is also how `LibraryDependenciesCache` sees it, because `computeLibrariesAndSdksUsedWithNoFilter` gets all dependent modules via the library info, so all dependent modules of all `LibraryEx`s are mushed together in the dependent module set of the "combining" `LibraryInfo`.

#### Solution

We need a solution which works with the caching of (transitive) dependents. We cannot just get all deduplicated libraries from `LibraryInfoCache`, because the libraries are added as they are requested and the cache in `IdeKotlinModuleDependentsProvider` is not cleared when a new deduplicated library is added to `LibraryInfoCache`. (Nor should it, because the deduplication cache is an implementation detail of `LibraryInfoCache`.)

Instead, we can use `LibraryUsageIndex` which `LibraryDependenciesCache` already relies on to find the dependent modules for some `LibraryInfo`, with full support for deduplicated libraries. This information is stable because it relies on module order entries.

Given that `LibraryUsageIndex` is properly cleared on project structure changes, as is `IdeKotlinModuleDependentsProvider`'s cache, there should be no caching issues with this approach.

#### Selective `LibraryEx` invalidation is messy

Another possible solution for invalidation in the example above is to not invalidate `M1`'s session, but instead to only invalidate `M3`'s session. This would be legal because `L2` didn't change at all, and we just need some new session for `L3`. However, this comes with a few issues:

1. It breaks the simple consistency of *always* invalidating at least the session for which a modification event was published. The special handling of `LibraryInfo`s breaks conceptual boundaries in the Analysis API.
2. The session for `M1` does eventually need to be removed when the last `LibraryEx` has been removed/replaced, which increases complexity.
3. We would need to introduce special handling for anchor libraries (if still applicable), because a library info is removed when its anchor library is removed/replaced. So we wouldn't be able to keep `M1`'s session in that case, because we need to lose the reference to the invalid `LibraryInfo`.
4. We'd need to implement some way to get the dependents of specific `LibraryEx`s like `L3` via the `LibraryInfo`/`KtLibraryModule` interface, which does not fit into the conceptual model of `LibraryInfo` as a combined view on a collection of `LibraryEx`.

^KTIJ-26376 fixed

GitOrigin-RevId: 72baaff8d9e8c62d40bd7b6be4a3cb1a60a1c8ad