IDEA-327562 Move setBorder to TreeTableTree.updateUI
It was called from the constructor, but it gets reset by updateUI
to whatever is specified in the UI Defaults, causing painting artifacts.
We can get around it by either using a proper non-UIResource
empty border or by resetting it ourselves in updateUI. Since we already have
an updateUI implementation here, the latter way makes more sense.

GitOrigin-RevId: 266c38af2e992fceedc5eb02db020ad8e9c3ddc0