Refactor light edit frame initialization code and menu bar helper logic
This commit refactors the light edit frame initialization code in 'LightEditFrameWrapper.kt', replacing 'runBlockingModalWithRawProgressReporter' with 'runWithModalProgressBlocking'. Additionally, it optimizes the asynchronous initialization of 'FileEditorManager'.
For 'IdeMenuBarHelper.kt', it removes the redundancy of setting the "always-visible" flag for main menu items within 'expandMainActionGroup'. This flag setting operation is now extracted and integrated with 'getMainMenuActionGroup', reducing code repetitions and potentially improving readability. These changes aim to simplify the codebase and increase efficiency.

GitOrigin-RevId: d359d8a85f2ff7d541e1381eee6800b58646e257