IDEA-319881, EA-819606 Use Project View renderers to create dragged image
These slow ops were caused by trying to reinvent the wheel and doing
custom rendering to create a dragged image, which involved messing around
with PSI and stuff.

Solve by simply delegating this job to the project view renderers
(and a simple JLabel to add that ... and N more files... message).

The result looks a bit different from the original code, but it's probably
a good thing, as it now looks exactly what is being dragged, which
is likely the expected behavior. The only issue is that now it also
includes inplace comments (Details in Tree View), which look out
of place there, but we'll deal with it shortly.

GitOrigin-RevId: 114ee862298fea21603b4adfd409aca8c381bf57