Fix concurrency issue in ModelBuilderService parameters
Before this change, there is only a single instance of
ModelBuilderContext per service instance and that is used to contain
the model builder parameters. But multiple `buildAll` actions can  be
run concurrently resulting in model builders getting the paramaters for
other model builders. This change moves the parameter out of the
ModelBuilderContext and to the method parameters to avoid this issue.

closes https://github.com/JetBrains/intellij-community/pull/2535

GitOrigin-RevId: 7ce07c98ab1799f72cae69b14dd649309f20056a