Fix concurrency issue in ModelBuilderService parameters
Before this change, there is only a single instance of
ModelBuilderContext per service instance and that is used to contain
the model builder parameters. But multiple `buildAll` actions can  be
run concurrently resulting in model builders getting the paramaters for
other model builders. This change moves the parameter out of the
ModelBuilderContext and to the method parameters to avoid this issue.

closes https://github.com/JetBrains/intellij-community/pull/2535

(cherry pick of commit 590bf9216c74e686dcec407fd24e65f6d014d5f5)

Change-Id: I590bf9216c74e686dcec407fd24e65f6d014d5f5