CPP-34495: ignore hardlinks in tar archives
tar archives might contain hardlinks which are considered
as files though this is wrong as they are not. their size
is 0 bytes, and they should be handled properly. since we
never expected hardlinks so far, we can ignore them at
unpacking state. the easiest way to get archive with hardlinks
is to tar some file twice and the second item will turn into
a hardlink.

GitOrigin-RevId: 00f527ada8443c99d7adde20513846aa4c60d8bc