IDEA-319881, EA-819606 Don't paint grayed text for dragged image
A bit of an ugly hack here, but it works, it's harmless enough as it's
module-local, and there doesn't seem to be an easy elegant way around it.

To avoid stuff like inplace comments (Details in Tree View) added to
the dragged image, we temporarily disable painting grayed out text.
Unfortunately, there is no easy way to tell inplace comments apart
from the rest of the presentation. In fact, NodeRenderer does
exactly the same thing, stopping at the first grayed out fragment
to determine the "main" searchable part of the node's presentation.

It's not clear whether it's actually needed to stop at the first grayed
out fragment, or we could just skip it and continue. But to mimic
NodeRenderer's behavior seems like the safest bet here, as
we generally don't want anything but the main part in the dragged image.

GitOrigin-RevId: 7cbcdda5dc50262127aef3d1b7d2ed33a21d4315