[terminal] IDEA-330183 Show terminal output in a single editor
It is a large refactoring with a lot of renames and architectural changes. These changes are needed to show the command outputs in the single editor. Also, I try to make the logic separated from the UI as much as possible.
Part of the previous functionality is supported after these changes:
- text highlights in the output
- alternate buffer support
- visual output block presentation
All of that will be fixed in the next commits.

GitOrigin-RevId: 97a63422d1053bf0afa66b94a2cea8530eab7bd8