IDEA-320010, EA-677913 Fix ScopeViewPane.select slow op
Caused by NamedScopeFilter doing its stuff when accept() is called.

Fix by moving this part to SelectInProjectViewImpl. While we're at it,
reduce raw PSI use in the EDT by using a smart pointer here.
It's also needed because once we enter the coroutine world,
all kinds of suspending and read actions are going to happen,
and there's no guarantee that PSI is not invalidated in the process.
So create a pointer as early as possible and pass it around.
It was eventually created anyway for the visitor, we're just
doing it earlier now, so that's a good thing with no trade-offs.

GitOrigin-RevId: ca3d594829901a5e2703e793b892b3efd7dcab7a