IDEA-326708 Log saving bounds of dialogs, add more monitor details
Besides specifying which monitor is the default, we also
now log which monitor is the IDE frame considers its own.
This can be tricky if it spans several monitors, but we log
what we're interested in: what ideFrame.getGraphicsConfiguration()
returns, as that's what the Dimension Service is using
to compose its keys.

GitOrigin-RevId: d2260c69bb44b547079a236ac9587dedec4ad26a