IDEA-328234 vcs: add icon & description to "Copy Revision Number" in annotations
CopyRevisionNumberAction gets its text, description, and icon via
<action> tag in VcsActions.xml and corresponding keys in
ActionsBundle.properties.  This action is available from tab "Log" of
tool window "Git" and similar UIs.

A similar action CopyRevisionNumberFromAnnotateAction has the same text,
which for historical reasons lives in a separate VcsActions.properties,
but doesn't have a description and an icon.

Add a description and an icon to CopyRevisionNumberFromAnnotateAction,
same as CopyRevisionNumberAction already has.