[platform] provide a way to specify a URL for bug reports in ExternalProductResourceUrls (IJPL-204)
Actually, it's not necessary to use YouTrack for bug reporting, so the introduced property is named 'bugReportUrl'. IDEs produced by JetBrains are supposed to use YouTrack, so a basic implementation was added to BaseJetBrainsExternalProductResourceUrls, and it's enough to specify only YouTrack Project ID in its implementations.

IntelliJ IDEA Ultimate has been migrated to use the new property.

GitOrigin-RevId: eb062526b26a22cfa344a03cebb3cc9272d1c395