[Kotlin] Fix exception reported when removing annotation from expect class
Because fix applied from file with actual declaration, but another file with expect
declaration is modified.

The following exception was reported:

```
Intention action class org.jetbrains.kotlin.idea.inspections.RemoveAnnotationFix
(Remove annotation from expect declaration) refers to the element from another
source file. It's likely that it's going to modify a file not opened in the editor,
so default preview strategy won't work. Also, if another file is modified,
getElementToMakeWritable() must be properly implemented to denote the actual file
to be modified.
```

^KTIJ-26633

GitOrigin-RevId: 0d173053550ac0583bfe1f167aad03506c1af375