improve JavaDoc: ProblemHighlightType.GENERIC_ERROR_OR_WARNING: add more code links to related entities
Previously JavaDoc of GENERIC_ERROR_OR_WARNING only mentioned `com.intellij.codeHighlighting.HighlightDisplayLevel`.
However, the calculation of the final highlighting style is not only determined by the highlighting level. It also takes into account "editorAttributes" (each inspection can be provided in "plugin.xml" and user can change it in the inspection settings).

Initially, I wanted to roughly describe the logic behind calculation of the final "TextAttributes" in the JavaDoc. But in the end, I decided it would be too much for the JavaDoc of GENERIC_ERROR_OR_WARNING and if needed, should be done somewhere else.

GitOrigin-RevId: a35d0b6419ebe9c1d4aac44a50df9596fbe3fb79