IDEA-316483 Use OnePixelSplitter in TextEditorWithPreview
This fixes the usability issue of the divider being too hard to target
with the mouse, as OnePixelSplitter uses a customized divider
implementation that uses the glass pane to enhance the mouse
grab zone.

Implementation note: for the old UI we still set the divider width
to 2, and OnePixelSplitter seems to support that just fine despite its name.
Using a 1-pixel splitter doesn't look very well due to lower theme contrast.

GitOrigin-RevId: 604ea64887c54fef0a3c82fe01c8564457e15fee