IDEA-329783 ProcessPopup: setLocateWithinScreenBounds(false)
Turns out the default is true, and it's not what users expect for this
particular popup, as it makes perfect sense to place it on a secondary display.

Also, it's the way it always worked until we delegated dimension service
stuff to AbstractPopup in IDEA-318686 (ProcessPopup used to have its
own save/restore logic before that), so that change was technically
a regression.

GitOrigin-RevId: 5d45cdb9b8dc3832683857629ca93ce8663dba2f