IDEA-328176 Add move listener for windowed tool windows
Well, this one is surely more complicated than it would seem.

The problem is that the windowed decorator is an instance of
a private class, MyJFrame in FrameWrapper.kt. So while we
can easily add a component listener to it, we can't add
an apply() method to it, nor can we find it in a manner
similar to how FloatingDecorator was located previously,
with SwingUtilities.getAncestorOfClass.

To make it work, introduce a new internal interface,
ToolWindowExternalDecorator, that contains the apply()
method for both windowed and floating decorators.
To locate it, we introduce a client property, which we
set on the actual decorator component, be it FloatingDecorator
or that window. This property contains a reference
to the actual implementation, which now doesn't have
to be a Swing component, so we just subclass
FrameWrapper to implement this interface for windowed
decorators. For floating decorators it's still the decorator
itself that implements it.

Moreover, we also need to add a method to check
the actual type, because at some points in a tool window
life cycle it's actually possible that its type doesn't
match its decorator, while the type is being updated.
For example, when switching from floating to windowed,
at some point the type is already windowed, but the
floating decorator is still there. So we check the decorator's
type and only call apply() if the type matches. Otherwise,
its the job of the code that changes the type to properly
apply the updated window info.

GitOrigin-RevId: 303d5f1b0756f2c8cce5ab3da7de3a421b22dc2e