CWM-8468 Fix Quick Definition popup is shown on wrong client
The reason popup wash shown on a wrong client is that action preserved the latest popup and updated it (instead of creating a new one) if it was still visible.
The action instance is the same for all clients, so the popup reference was also shared between them all.

To fix the problem, I've created a per-client service which is responsible for showing and updating implementations popup. The reference to an active popup will be different for every client

GitOrigin-RevId: 6e2527c05a9c835014676599e154ca8c9185007c