Fix search scope for `KtLibrarySourceModules` in scope provider.
Without this fix, when creating a search scope for a library,
the contents of that library itself was not included in the search
scope.

That causes issue for the use in evaluate expression. When
performing evaluate expression at a breakpoint in a binary
library dependency, class files for the library itself
could not be found.

^KTIJ-26694 Fixed

GitOrigin-RevId: 3df2d8a02dbb4733bd96836b17d115c9451e27c3
(cherry picked from commit b81924e8cd09c7a73d00c9b7b84017dbebf0f742)
Change-Id: Iadab1e568e541346e94348775945fe2eb9f132ff
Bug: 296377820