CWM-8468 Support pin action update in ImplementationViewComponent
Several different actions are using the same instance of ImplementationPopupManager which preservers the latest shown popup with ImplementationViewComponent.
Even though the ImplementationViewComponent support content updates without component recreation, it doesn't support toolbar updates. So, pin (show in tool window)  action is either always shown or always hidden which prevents actions with different pin (`couldPinPopup`) from updating popup via ImplementationPopupManager.

Instead of passing processor to consumer and adding (or not adding) the action in toolbar depending on the value, I've added the new method `setShowInFindWindowProcessor` which updates the processor.
The action is updating its presentation based on the value of `myShowInFindWindowProcessor` (if null - the action will be invisible and disabled)

GitOrigin-RevId: 62e86dd6ce3bf30330ad429f25f198ec203d09ba