IDEA-329995 Fix ToolWindowDragHelper.canStartDragging false positive
Turned out harder than it seemed. The issue was caused by
ToolWindowDragHelper checking whether the dragging is
started on the right component. This was implemented using
getDeepestComponentAt on the tool window pane and its parent.
This didn't work because it didn't take possible lightweight popups
into account.

The obvious solution would be to simply start looking from
the topmost window parent, but it didn't work either because
the glass pane was mistakenly returned then, thanks to
this very ToolWindowDragHelper adding a drop target component
to it.

So we need to start looking from the layered pane instead.
But no, just using
ComponentUtil.getParentOfType(JLayeredPane::class.java, dragSourcePane)
didn't work either, because it would return the tool window pane itself
instead, as it also extends JLayeredPane (surprise!) and getParentOfType()
considers the object itself to be its own parent (double surprise!).

So to make it work we have to look for the topmost JLayeredPane parent
and start looking for the deepest component from there. This way
it works for stripe buttons and tool windows unless they're obscured
by something, just as we want it.

GitOrigin-RevId: 39194a6c2c100bf7f0f508474e19cdcd12372715