CPP-34359: Remove empty DEFAULT_PARAMETER key
Default theme's DEFAULT_PARAMETER should fallback to IDENTIFIER, just like in Darcula.
This fixes issues where an external highlighter returns a DEFAULT_PARAMETER highlighting, but a wrong highlighting is applied instead. This happens because looking up the key in the scheme finds the empty DEFAULT_PARAMETER value with null colors, thus we fallback to the other highlighter's colors, which may be incorrect.

GitOrigin-RevId: ea4fb9fccf4f164c677d1eb5ce09638c0c402b49