IDEA-330161 Don't dispatch key events to popup when a modal dialog is open
Due to the way IdeEventQueue and IdePopupManager are designed,
they bypass normal key event handling and send events directly to popups
if there are any. But since we now allow modal dialogs to be opened
from popups, as long as such a dialog is displayed, it should handle
all events instead. Otherwise, a popup may try to close, e.g. on Esc,
and since we don't allow it, it ends up in a weird state, not really open,
not really closed.

GitOrigin-RevId: 792729781861364ba14b7221f066ecf1cae80a62