IDEA-326708 Move DialogWrapperPeerImpl trace logging to reshape()
This is just to be 100% sure that we log every single API call, as there
are different ones (setSize, setLocation, setBounds) and they all
delegate to this one internally. So even though it's deprecated,
it looks like the perfect place for logging, otherwise we may miss
some direct reshape() calls from legacy code or JDK internals.

GitOrigin-RevId: 4db0beee1b878bf1718a5af87e19794d727c31af