[gradle] GradleAttachSourcesProvider: try to attach sources from Gradle User Home before downloading; verify jar header
In the previous revision of GradleAttachSourcesProvider, we tried to download sources via Gradle and only then search for them in the cache. The cache lookup after task execution may be omitted since we know the exact location of the downloaded sources after the Gradle task finish. The cache lookup was moved before Gradle task execution to prevent Gradle for launching for local-available sources.

Validate header of the downloaded/local available Jar file.

GitOrigin-RevId: 843b84df121698b8b8f483d18302b07aafa7ba4e